--- a/.gitignore
+++ b/.gitignore
@@ -499,3 +499,4 @@ _temp/
 _output/
 API/stats/
 UI/Web/dist/
+/API.Tests/Extensions/Test Data/modified on run.txt