--- a/.gitignore
+++ b/.gitignore
@@ -500,3 +500,4 @@ _output/
 API/stats/
 UI/Web/dist/
 /API.Tests/Extensions/Test Data/modified on run.txt
+/API/covers/