--- a/API.Benchmark/ArchiveSerivceBenchmark.cs
+++ b/API.Benchmark/ArchiveSerivceBenchmark.cs
@@ -1,0 +1,8 @@
+﻿namespace API.Benchmark
+{
+    public class ArchiveSerivceBenchmark
+    {
+        // Benchmark to test default GetNumberOfPages from archive
+        // vs a new method where I try to open the archive and return said stream
+    }
+}