--- a/API.Benchmark/Program.cs
+++ b/API.Benchmark/Program.cs
@@ -12,7 +12,8 @@ namespace API.Benchmark
     {
         static void Main(string[] args)
         {
-            BenchmarkRunner.Run<ParseScannedFilesBenchmarks>();
+            //BenchmarkRunner.Run<ParseScannedFilesBenchmarks>();
+            BenchmarkRunner.Run<TestBenchmark>();
         }
     }
 }