--- a/API.Tests/Parser/MangaParserTests.cs
+++ b/API.Tests/Parser/MangaParserTests.cs
@@ -236,6 +236,7 @@ namespace API.Tests.Parser
         [InlineData("Kedouin Makoto - Corpse Party Musume, Chapter 09.cbz", "9")]
         [InlineData("Hentai Ouji to Warawanai Neko. - Vol. 06 Ch. 034.5", "34.5")]
         [InlineData("Kimi no Koto ga Daidaidaidaidaisuki na 100-nin no Kanojo Chapter 1-10", "1-10")]
+        [InlineData("Deku_&_Bakugo_-_Rising_v1_c1.1.cbz", "1.1")]
         public void ParseChaptersTest(string filename, string expected)
         {
             Assert.Equal(expected, API.Parser.Parser.ParseChapter(filename));