--- a/API/DTOs/Downloads/DownloadBookmarkDto.cs
+++ b/API/DTOs/Downloads/DownloadBookmarkDto.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using API.DTOs.Reader;
 
 namespace API.DTOs.Downloads
 {