… API.Tests/Extensions/Test Data/modified on run.txt
… API.Tests/Services/ArchiveServiceTests.cs
… API.Tests/Services/Test Data/ImageService/cover.expected.jpg
… API/Comparators/ChapterSortComparer.cs
… API/Comparators/NaturalSortComparer.cs
… API/Controllers/BookController.cs
… API/Controllers/CollectionController.cs
… API/Controllers/DownloadController.cs
… API/Controllers/ImageController.cs
… API/Controllers/LibraryController.cs
… API/Controllers/OPDSController.cs
… API/Controllers/PluginController.cs
… API/Controllers/ReaderController.cs
… API/Controllers/ReadingListController.cs
… API/Controllers/SeriesController.cs
… API/Controllers/SettingsController.cs
… API/Controllers/UploadController.cs
… API/Controllers/UsersController.cs
… API/DTOs/Account/LoginDto.cs
… API/DTOs/Account/ResetPasswordDto.cs
… API/DTOs/ImageDto.cs
… API/DTOs/InProgressChapterDto.cs
… API/DTOs/Reader/BookInfoDto.cs
… API/DTOs/Reader/BookmarkDto.cs
… API/DTOs/Reader/ChapterInfoDto.cs
… API/DTOs/Reader/IChapterInfoDto.cs
… API/DTOs/Reader/MarkMultipleSeriesAsReadDto.cs
… API/DTOs/Reader/MarkReadDto.cs
… API/DTOs/Reader/MarkVolumeReadDto.cs
… API/DTOs/Reader/MarkVolumesReadDto.cs